Fix : user_id 타입 수정 Post 속성 수정

+++ TigerCave/src/main/java/com/sparta/tigercave/entity/PostLike.java
@@ -4,8 +4,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.math.BigInteger;
-
 @Entity
 @Getter
 @NoArgsConstructor
@@ -14,7 +12,7 @@ public class PostLike {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "post_like_id")
-    private BigInteger id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id" )
