+++ TigerCave/src/main/java/com/sparta/tigercave/repository/PostLikeRepository.java
@@ -5,10 +5,8 @@ import com.sparta.tigercave.entity.PostLike;
 import com.sparta.tigercave.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 import java.util.List;
-import java.util.Optional;
 
-public interface PostLikeRepository extends JpaRepository<PostLike, BigInteger> {
+public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
     List<PostLike> findByUserAndPost(User user,Post post);
 }
