+++ TigerCave/src/main/java/com/sparta/tigercave/repository/UsersRepository.java
@@ -3,9 +3,8 @@ package com.sparta.tigercave.repository;
 import com.sparta.tigercave.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<User, BigInteger> {
+public interface UsersRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 }
